Drop toolbox entirely
