--- toolbox/util/BUILD.bazel
@@ -1,36 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = [
-        "commonUtils.go",
-        "dependency.go",
-        "gcsClient.go",
-        "githubClient.go",
-        "githubUtils.go",
-    ],
-    importpath = "istio.io/test-infra/toolbox/util",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//vendor/cloud.google.com/go/storage:go_default_library",
-        "//vendor/github.com/golang/glog:go_default_library",
-        "//vendor/github.com/google/go-github/github:go_default_library",
-        "//vendor/github.com/hashicorp/go-multierror:go_default_library",
-        "//vendor/golang.org/x/oauth2:go_default_library",
-        "//vendor/google.golang.org/api/iterator:go_default_library",
-    ],
-)
-
-filegroup(
-    name = "package-srcs",
-    srcs = glob(["**"]),
-    tags = ["automanaged"],
-    visibility = ["//visibility:private"],
-)
-
-filegroup(
-    name = "all-srcs",
-    srcs = [":package-srcs"],
-    tags = ["automanaged"],
-    visibility = ["//visibility:public"],
-)
